Update Clippy with auto-scaling

+++ Cogs/Clippy.py
@@ -39,6 +39,7 @@ class Clippy:
                 # add the line to the lines array
                 lines.append(line)
         return lines
+
     
     @commands.command()
     async def clippy(self, ctx, *,text: str = ""):
@@ -65,22 +66,32 @@ class Clippy:
         if not len(text):
             text = random.choice(clippy_errors)
 
-        font = ImageFont.truetype('fonts/comic.ttf', size=20)
-        #340 is the width we want to set the image width to
-        lines = self.text_wrap(text, font, 340)
-        line_height = font.getsize('hg')[1]
-        (x, y) = (25, 20)
-        color = 'rgb(0, 0, 0)' # black color
-        text_size = draw.textsize(text, font=font)
+        for xs in range(30, 2, -1):
+            font = ImageFont.truetype('fonts/comic.ttf', size=xs)
+            #340 is the width we want to set the image width to
+            lines = self.text_wrap(text, font, 340)
+            line_height = font.getsize('hg')[1]
+            (x, y) = (25, 20)
+            color = 'rgb(0, 0, 0)' # black color
+            text_size = draw.textsize(text, font=font)
 
-        for line in lines:
-            text_size = draw.textsize(line, font=font)
-            image_x = (image_width /2 ) - (text_size[0]/2)
-            draw.text((image_x, y), line, fill=color, font=font)
-            y = y + line_height
+            for line in lines:
+                text_size = draw.textsize(line, font=font)
+                image_x = (image_width /2 ) - (text_size[0]/2)
+                draw.text((image_x, y), line, fill=color, font=font)
+                y = y + line_height
+            if y < 182: # Check if the text overlaps. 182 was found by trial and error.
+                image = Image.open('images/clippy.png')
+                draw = ImageDraw.Draw(image)
+                (x, y) = (25, 20)
+                for line in lines:
+                    text_size = draw.textsize(line, font=font)
+                    image_x = (image_width /2 ) - (text_size[0]/2)
+                    draw.text((image_x, y), line, fill=color, font=font)
+                    y = y + line_height
+                image.save('images/clippynow.png')
+                await ctx.send(file=discord.File(fp='images/clippynow.png'))
 
-        image.save('images/clippynow.png')
-        await ctx.send(file=discord.File(fp='images/clippynow.png'))
-
-        # Remove the png
-        os.remove("images/clippynow.png")
+                # Remove the png
+                os.remove("images/clippynow.png")
+                break
